Finally working. (Usage of Box<T>)
The problem was because I stored the graphs on the stack and not on the heap. I imagine the problem was coming from graphs that needed many recursive calls that stores a lot of graphs on the stack.

Anyway, the problem is now solved, understanding the problem in details is a job for the future me.